refactor: actualizar manejo de UserId a Long, mejorar integración de TokenManager y corregir imports en modelos y repositorios de usuario
- Cambiar tipo de usuarioId de Int a Long en TokenManager y modelos relacionados
- Ajustar lógica de sincronización y obtención de userId en ViewModels y repositorios
- Mejorar integración de TokenManager en MisPedidosActivity y MisPedidosViewModel
- Corregir imports y convenciones de nombres en modelos y repositorios de usuario
- Actualizar dependencias y configuración de plugins en build.gradle.kts y libs.versions.toml para Kotlin 2.1 y Compose

+++ app/src/main/java/com/example/mimascota/util/TokenManager.kt
@@ -47,7 +47,7 @@ object TokenManager {
         val json = gson.toJson(usuario)
         prefs.edit().apply {
             putString(KEY_USUARIO, json)
-            putInt(KEY_USER_ID, usuario.usuarioId)
+            putLong(KEY_USER_ID, usuario.usuarioId.toLong())
             apply()
         }
     }
@@ -71,8 +71,8 @@ object TokenManager {
     /**
      * Obtener ID del usuario
      */
-    fun getUserId(): Int {
-        return prefs.getInt(KEY_USER_ID, -1)
+    fun getUserId(): Long {
+        return prefs.getLong(KEY_USER_ID, -1L)
     }
 
     /**
@@ -143,4 +143,3 @@ object TokenManager {
         }
     }
 }
-
